# docs/contracts/generated/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lunar Analyst API",
    "description": "Typed-contract-first FastAPI stub",
    "version": "1.0.0-stage1"
  },
  "paths": {
...
    "/api/v1/notebook/sessions": {
      "post": {
        "tags": [
          "v1"
        ],
        "summary": "Create Notebook Session",
        "operationId": "create_notebook_session_api_v1_notebook_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotebookSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotebookSession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notebook/sessions/{session_id}": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Get Notebook Session",
        "operationId": "get_notebook_session_api_v1_notebook_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotebookSession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateNotebookSessionRequest": {
        "properties": {
          "client_name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Client Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "client_name"
        ],
        "title": "CreateNotebookSessionRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotebookSession": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "api_token": {
            "type": "string",
            "title": "Api Token"
          },
          "client_name": {
            "type": "string",
            "title": "Client Name"
          },
          "created_at_utc": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}-\\d{2}-\\d{2}$",
            "title": "Created At Utc"
          },
          "last_seen_at_utc": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}-\\d{2}-\\d{2}$",
            "title": "Last Seen At Utc"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "api_token",
          "client_name",
          "created_at_utc",
          "last_seen_at_utc"
        ],
        "title": "NotebookSession"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}